Moved TickCalculator functionality into grid and deleted TickCalculator
In order to keep related methods and data together, the TickCalculator functionality was moved into the Grid class, which is where it will be used. The tests for the TickCalculator were also moved into the GridTest file and the calls were refactored accordingly. The Grid can now produce a new Grid object.