Changed contract for Grid class
In doing so it changed the requirements for many classes that interact with it, so these all had to be changed too. The Grid class now is constructed with and gives out a List of Cells, rather than an array of arrays. Tests were moved around and re-written in the process, as it turned out that behaviour was moving from one class to another. Deficencies in the test coverage were also uncovered.