Moved the adding of dead cells out of the grid class
To simplify the grid class constructor, the adding of dead cells was removed from the grid class, so it now takes a list of all the cells rather than sometimes taking a list of live cells and sometimes taking the full list. This logic is currently in ConwaysApp but hopefully will move into an as-yet-undiscovered class.