Changed Cell aliveness to an Enum rather than a boolean
This reflects the domain better and allows for the expansion into further types of life if necessary (e.g. zombie). The Grid Calculator has also been updated to be able to ask for cell neighbours of a particular type, not just for alive neighbours. During this refactor the boolean array of arrays was also completed eradicated 🎉