Removed the HTTPExchange dependency from some classes (with help)
Pulled the HttpExchange out of some classes so that they are more testable. Also discovered that the put does actually work, I'd just used the testing library incorrectly and had made the Postman request an Https rather than an Http and it was waiting for a TLS handshake and never getting one. Lots of small fixes that were very effective.