Fixed PUT bug and added run command to README
When PUT was called with the seed name it wasn't deleting the seed name but it was adding the new name. Fixed it so that it returns a 405 - method not allowed and a message that you can't replace the seed name.
Added the run command to the readme as realised it wasn't there when I went to copy it from there.
 Please enter the commit message for your changes. Lines starting